+++ CHANGELOG.md
@@ -1,19 +1,37 @@
-## 1.1.0 (2021-11-26)
+### 1.2.1-beta.2 (2022-10-06) 
+
+
+### Bug Fixes
+
+* refine logs ([5aa3ef8](https://github.com/abstracter-io/atomic-release/commit/5aa3ef89bd243c9c580fba8f5f76f2d466520656))
+
+
+### 1.2.1-beta.1 (2022-10-06)
 
 
 ### Features
 
-* add release implementation for trunk based development ([23f4a31](https://github.com/abstracter-io/atomic-release/commit/23f4a31383f0fd4e3e88132149dfb55629b4da12))
+* add a default undo/cleanup ([589f74e](https://github.com/abstracter-io/atomic-release/commit/589f74eedf467919beae126b56671490ffaf70c7))
+* refine log for succeeded/failed release ([ebc5a48](https://github.com/abstracter-io/atomic-release/commit/ebc5a4804228d197fed01a8e5549cca598ef1976))
 
 
-## 1.0.0 (2021-11-26)
+### Bug Fixes
+
+* **timer:** elapsed time ms calculation ([b822f39](https://github.com/abstracter-io/atomic-release/commit/b822f39d10e8d194f0b35476fef0f67100fd4f28))
 
 
-### ⚠ BREAKING CHANGES
+### 1.2.1-beta.0 (2022-10-06)
 
-* initial version
+
+### Bug Fixes
+
+* **trunk-release:** use 'HEAD' implicitly ([87d1261](https://github.com/abstracter-io/atomic-release/commit/87d1261e9eb842c1a4745b59b32810e2afcdad6d))
+
+
+## 1.2.0 (2022-10-06)
+
 
 ### Features
 
-* initial version ([35015c5](https://github.com/abstracter-io/atomic-release/commit/35015c5c6cde18050139a6f5db5a8062fb993fb1))
+* version 1.2.0 ([058cebd](https://github.com/abstracter-io/atomic-release/commit/058cebd161eaa09f624695e0483b398681109d4c))